+++ tests/clients/test_ngsi_ld_subscription.py
@@ -3,25 +3,24 @@ Test the endpoint for subscription related task of NGSI-LD for ContextBrokerClie
 """
 import json
 import time
+import urllib.parse
 from unittest import TestCase
-from pydantic import ValidationError
 import threading
 from paho.mqtt.enums import CallbackAPIVersion
 import paho.mqtt.client as mqtt
+from requests import RequestException
 from filip.clients.ngsi_ld.cb import ContextBrokerLDClient
 from filip.models.base import FiwareLDHeader
 from filip.models.ngsi_ld.context import \
-    ContextProperty, \
     NamedContextProperty, \
-    ContextLDEntity
+    ContextLDEntity, ActionTypeLD
 from filip.models.ngsi_ld.subscriptions import \
     Endpoint, \
     NotificationParams, \
-    Subscription
-from filip.utils.cleanup import clear_all, clean_test
+    SubscriptionLD
 from tests.config import settings
-from random import randint
-from pydantic import AnyUrl
+from filip.utils.cleanup import clear_context_broker_ld
+
 
 class TestSubscriptions(TestCase):
     """
@@ -37,48 +36,23 @@ class TestSubscriptions(TestCase):
         self.fiware_header = FiwareLDHeader(ngsild_tenant=settings.FIWARE_SERVICE)
         self.cb_client = ContextBrokerLDClient(fiware_header=self.fiware_header,
                                                url=settings.LD_CB_URL)
-        # initial tenant
-        self.cb_client.post_entity(ContextLDEntity(id="Dummy:1", type="Dummy"),
-                                   update=True)
-        self.cb_client.delete_entity_by_id("Dummy:1")
-        # self.mqtt_url = "mqtt://test.de:1883"
-        # self.mqtt_topic = '/filip/testing'
-        # self.notification =  {
-        #     "attributes": ["filling", "controlledAsset"],
-        #     "format": "keyValues",
-        #     "endpoint": {
-        #     "uri": "http://test:1234/subscription/low-stock-farm001-ngsild",
-        #     "accept": "application/json"
-        #     }
-        # }
-        self.cb_client = ContextBrokerLDClient()
+        clear_context_broker_ld(cb_ld_client=self.cb_client)
+
         self.mqtt_topic = ''.join([settings.FIWARE_SERVICE,
                                    settings.FIWARE_SERVICEPATH])
-        self.MQTT_BROKER_URL_INTERNAL = "mqtt://mosquitto:1883"
-        self.MQTT_BROKER_URL_EXPOSED = "mqtt://localhost:1883"
         self.endpoint_mqtt = Endpoint(**{
-            "uri": "mqtt://my.host.org:1883/my/test/topic",
-            "accept": "application/json",  # TODO check whether it works
+            "uri": str(settings.LD_MQTT_BROKER_URL) + "/my/test/topic",
+            "accept": "application/json",
         })
-        self.cb_client = ContextBrokerLDClient(url=settings.LD_CB_URL, fiware_header=self.fiware_header)
         self.endpoint_http = Endpoint(**{
-            "uri": "http://137.226.248.246:1027/ngsi-ld/v1/subscriptions",
+            "uri": urllib.parse.urljoin(str(settings.LD_CB_URL),
+                                        "/ngsi-ld/v1/subscriptions"),
             "accept": "application/json"
-        }
-                                      )
-        self.cleanup()
+        })
 
     def tearDown(self) -> None:
-        self.cleanup()
-    
-    def cleanup(self):
-        """
-        Cleanup test subscriptions
-        """
-        sub_list = self.cb_client.get_subscription_list()
-        for sub in sub_list:
-            if sub.id.startswith('urn:ngsi-ld:Subscription:test_sub'):
-                self.cb_client.delete_subscription(sub.id)
+        clear_context_broker_ld(cb_ld_client=self.cb_client)
+        self.cb_client.close()
 
     def test_post_subscription_http(self):
         """
@@ -93,7 +67,7 @@ class TestSubscriptions(TestCase):
         attr_id = "attr"
         id = "urn:ngsi-ld:Subscription:" + "test_sub0"
         notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)
-        sub = Subscription(id=id, notification=notification_param, entities=[{"type": "Room"}])
+        sub = SubscriptionLD(id=id, notification=notification_param, entities=[{"type": "Room"}])
         self.cb_client.post_subscription(sub)
         sub_list = [x for x in self.cb_client.get_subscription_list() 
                     if x.id == 'urn:ngsi-ld:Subscription:test_sub0']
@@ -128,7 +102,7 @@ class TestSubscriptions(TestCase):
         attr_id = "attr"
         id = "urn:ngsi-ld:Subscription:" + "test_sub0"
         notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)
-        sub = Subscription(id=id, notification=notification_param, entities=[{"type": "Room"}])
+        sub = SubscriptionLD(id=id, notification=notification_param, entities=[{"type": "Room"}])
         self.cb_client.post_subscription(sub)
         sub_get = self.cb_client.get_subscription(subscription_id=id)
         self.assertEqual(sub.entities, sub_get.entities)
@@ -151,7 +125,7 @@ class TestSubscriptions(TestCase):
             attr_id = "attr" + str(i)
             id = "urn:ngsi-ld:Subscription:" + "test_sub" + str(i)
             notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)
-            sub = Subscription(id=id, notification=notification_param, entities=[{"type": "Room"}])
+            sub = SubscriptionLD(id=id, notification=notification_param, entities=[{"type": "Room"}])
             sub_post_list.append(sub)
             self.cb_client.post_subscription(sub)
 
@@ -164,7 +138,6 @@ class TestSubscriptions(TestCase):
         sub_limit = 5
         sub_list2 = self.cb_client.get_subscription_list(limit=sub_limit)
         self.assertEqual(len(sub_list2), sub_limit)
-
 
     def test_delete_subscription(self):
         """
@@ -181,10 +154,10 @@ class TestSubscriptions(TestCase):
             notification_param = NotificationParams(
                 attributes=[attr_id], endpoint=self.endpoint_http)
             id = "urn:ngsi-ld:Subscription:" + "test_sub" + str(i)
-            sub = Subscription(id=id,
-                               notification=notification_param,
-                               entities=[{"type": "Room"}]
-                               )
+            sub = SubscriptionLD(id=id,
+                                 notification=notification_param,
+                                 entities=[{"type": "Room"}]
+                                 )
 
             if i == 0: 
                 del_sub = sub
@@ -220,26 +193,24 @@ class TestSubscriptions(TestCase):
         attr_id = "attr"
         id = "urn:ngsi-ld:Subscription:" + "test_sub77"
         notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)
-        sub = Subscription(id=id, notification=notification_param, entities=[{"type": "Room"}])
+        sub = SubscriptionLD(id=id, notification=notification_param, entities=[{"type": "Room"}])
         self.cb_client.post_subscription(sub)
         sub_list = self.cb_client.get_subscription_list()
         self.assertEqual(len(sub_list), 1)
         print(self.endpoint_http.model_dump())
-        sub_changed = Subscription(id=id, notification=notification_param, entities=[{"type": "House"}])
+        sub_changed = SubscriptionLD(id=id, notification=notification_param, entities=[{"type": "House"}])
         self.cb_client.update_subscription(sub_changed)
         u_sub= self.cb_client.get_subscription(subscription_id=id)
         self.assertNotEqual(u_sub,sub_list[0])
         self.maxDiff = None
-        self.assertDictEqual(sub_changed.model_dump(),u_sub.model_dump())
-        sub_list = self.cb_client.get_subscription_list()
-        self.assertEqual(u_sub.model_dump(),sub_list[0])
-        non_sub = Subscription(id="urn:ngsi-ld:Subscription:nonexist", 
-                                notification=notification_param, 
-                                entities=[{"type":"house"}])
+        self.assertDictEqual(sub_changed.model_dump(),
+                             u_sub.model_dump())
+        non_sub = SubscriptionLD(id="urn:ngsi-ld:Subscription:nonexist",
+                                 notification=notification_param,
+                                 entities=[{"type":"house"}])
         with self.assertRaises(Exception):
             self.cb_client.update_subscription(non_sub)
-        #Try to patch more than one subscription at once.
-        # TODO
+
 
 class TestSubsCheckBroker(TestCase):
     """
@@ -257,10 +228,14 @@ class TestSubsCheckBroker(TestCase):
         for sub in sub_list:
             if sub.id.startswith('urn:ngsi-ld:Subscription:test_sub'):
                 self.cb_client.delete_subscription(sub.id)
-        entity_list = self.cb_client.get_entity_list()
-        for entity in entity_list:
-            if entity.id.startswith('urn:ngsi-ld:Entity:test_entity'):
-                self.cb_client.delete_entity_by_id(entity_id=entity.id)
+        try:
+            entity_list = True
+            while entity_list:
+                entity_list = self.cb_client.get_entity_list(limit=100)
+                self.cb_client.entity_batch_operation(action_type=ActionTypeLD.DELETE,
+                                                      entities=entity_list)
+        except RequestException:
+            pass
 
     def setUp(self) -> None:
         """
@@ -268,45 +243,43 @@ class TestSubsCheckBroker(TestCase):
         Returns:
             None
         """
-        self.MQTT_BROKER_URL_INTERNAL = "mqtt://mqtt-broker-ld:1883"
-        self.MQTT_BROKER_URL_INTERNAL = AnyUrl(self.MQTT_BROKER_URL_INTERNAL)
         self.entity_dict = {
-            'id':'urn:ngsi-ld:Entity:test_entity03',
-            'type':'Room',
-            'temperature':{
-                'type':'Property',
-                'value':30
+            'id': 'urn:ngsi-ld:Entity:test_entity03',
+            'type': 'Room',
+            'temperature': {
+                'type':  'Property',
+                'value': 30
             }
         }
         
         self.sub_dict = {
-            'description':'Test Subscription',
-            'id':'urn:ngsi-ld:Subscription:test_sub25',
-            'type':'Subscription',
-            'entities':[
+            'description': 'Test Subscription',
+            'id': 'urn:ngsi-ld:Subscription:test_sub25',
+            'type': 'Subscription',
+            'entities': [
                 {
-                    'type':'Room'
+                    'type': 'Room'
                 }
             ],
-            'watchedAttributes':[
+            'watchedAttributes': [
                 'temperature'
             ],
-            'q':'temperature<30',
-            'notification':{
-                'attributes':[
+            'q': 'temperature<30',
+            'notification':  {
+                'attributes': [
                     'temperature'
                 ],
-                'format':'normalized',
-                'endpoint':{
-                    'uri':f'mqtt://'
-                          f'{settings.MQTT_BROKER_URL_INTERNAL.host}:'
-                          f'{settings.MQTT_BROKER_URL_INTERNAL.port}/my/test/topic', # change uri
-                    'Accept':'application/json'
+                'format': 'normalized',
+                'endpoint': {
+                    'uri':  f'mqtt://'  # change uri
+                          f'{settings.LD_MQTT_BROKER_URL_INTERNAL.host}:'
+                          f'{settings.LD_MQTT_BROKER_URL_INTERNAL.port}/my/test/topic',
+                    'Accept': 'application/json'
                 },
-                'notifierInfo':[
+                'notifierInfo': [
                     {
-                        "key":"MQTT-Version",
-                        "value":"mqtt5.0"
+                        "key": "MQTT-Version",
+                        "value": "mqtt5.0"
                     }
                 ]
             }
@@ -339,12 +312,12 @@ class TestSubsCheckBroker(TestCase):
 
         self.timeout = 5 # in seconds
         self.last_test_timeout = [True]
-        self.timeout_proc = threading.Timer(self.timeout,self.timeout_func)
-
+        self.timeout_proc = threading.Timer(self.timeout,
+                                            self.timeout_func)
 
     def tearDown(self) -> None:
         self.cleanup()
-
+        self.cb_client.close()
 
     def test_post_subscription_mqtt(self):
         """
@@ -364,12 +337,12 @@ class TestSubsCheckBroker(TestCase):
                              json.loads(msg.payload.decode())['body']['data'][0])
 
         self.mqtt_client.on_message = on_message
-        self.mqtt_client.connect(settings.MQTT_BROKER_URL.host,
-                                 settings.MQTT_BROKER_URL.port,
+        self.mqtt_client.connect(settings.LD_MQTT_BROKER_URL.host,
+                                 settings.LD_MQTT_BROKER_URL.port,
                                  60)
         self.mqtt_client.loop_start()
         #post subscription then start timer
-        self.cb_client.post_subscription(subscription=Subscription(**self.sub_dict))
+        self.cb_client.post_subscription(subscription=SubscriptionLD(**self.sub_dict))
         self.timeout_proc.start()
         #update entity to (ideally) get a notification
         self.cb_client.update_entity_attribute(entity_id='urn:ngsi-ld:Entity:test_entity03',
@@ -421,9 +394,11 @@ class TestSubsCheckBroker(TestCase):
 
         self.mqtt_client.on_message = on_message
         
-        self.mqtt_client.connect("localhost",1883,60)
+        self.mqtt_client.connect(settings.LD_MQTT_BROKER_URL.host,
+                                 settings.LD_MQTT_BROKER_URL.port,
+                                 60)
         self.mqtt_client.loop_start()
-        self.cb_client.post_subscription(subscription=Subscription(**self.sub_dict))
+        self.cb_client.post_subscription(subscription=SubscriptionLD(**self.sub_dict))
         self.timeout_proc.start()
 
         self.cb_client.update_entity_attribute(entity_id='urn:ngsi-ld:Entity:test_entity03',
@@ -433,13 +408,14 @@ class TestSubsCheckBroker(TestCase):
                                                 attr_name='temperature')
         while(self.timeout_proc.is_alive()):
             continue
-        self.assertTrue(self.last_test_timeout[0],"Operation timed out")
+        self.assertTrue(self.last_test_timeout[0],
+                        "Operation timed out")
 
         self.last_test_timeout = [True]
         self.timeout_proc = threading.Timer(self.timeout,self.timeout_func)
         
         self.sub_dict.update({'q':'temperature>30'})
-        self.cb_client.update_subscription(subscription=Subscription(**self.sub_dict))
+        self.cb_client.update_subscription(subscription=SubscriptionLD(**self.sub_dict))
         time.sleep(5)
         updated = self.cb_client.get_subscription(self.sub_dict['id'])
         self.assertEqual(updated.q,'temperature>30')
@@ -451,7 +427,8 @@ class TestSubsCheckBroker(TestCase):
                                             attr_name='temperature')
         while(self.timeout_proc.is_alive()):
             continue
-        self.assertTrue(self.last_test_timeout[0],"Operation timed out")
+        self.assertTrue(self.last_test_timeout[0],
+                        "Operation timed out")
         self.mqtt_client.loop_stop()
         self.mqtt_client.disconnect()