+++ tests/config.py
@@ -60,6 +60,17 @@ class TestSettings(BaseSettings):
                                                  'MQTT_BROKER_URL_INTERNAL',
                                                  'MQTT_URL_INTERNAL'))
 
+    LD_MQTT_BROKER_URL: AnyUrl = Field(default="mqtt://127.0.0.1:1883",
+                                       validation_alias=AliasChoices(
+                                             'LD_MQTT_BROKER_URL',
+                                             'LD_MQTT_URL',
+                                             'LD_MQTT_BROKER'))
+
+    LD_MQTT_BROKER_URL_INTERNAL: AnyUrl = Field(default="mqtt://mqtt-broker-ld:1883",
+                                                validation_alias=AliasChoices(
+                                                    'LD_MQTT_BROKER_URL_INTERNAL',
+                                                    'LD_MQTT_URL_INTERNAL'))
+
     # IF CI_JOB_ID is present it will always overwrite the service path
     CI_JOB_ID: Optional[str] = Field(default=None,
                                      validation_alias=AliasChoices('CI_JOB_ID'))